- reducing the hook of `DamageArea` also allowing compiler to optimize it more - remove cell spread limit of 255 - fix bug where `remove_if` is not properly erasing vector items - replace some hook with JUMP for better compatibility - Add PR `#1487`

+++ src/Misc/DynamicPatcher/Techno/GiftBox/GiftBoxFunctional.cpp
@@ -254,7 +254,7 @@ void GiftBox::Release(TechnoClass* pOwner, GiftBoxData& nData)
 
 		for (auto const& pTech : nOut) {
 			if (nData.RandomRange > 0) {
-				if (auto const pNewCell = GeneralUtils::GetCell(pCell, location, (size_t)(nData.RandomRange.Get()), nData.EmptyCell))
+				if (auto const pNewCell = GeneralUtils::GetCell(pCell, location, (short)(nData.RandomRange.Get()), nData.EmptyCell))
 					pCell = pNewCell;
 			}
